:bug: fix ranking of exam grades

+++ quddy/src/main/java/com/team_quddy/quddy/submit/domain/dto/SubmitResultDto.java
@@ -10,5 +10,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class SubmitResultDto {
     private Users users;
-    private Integer acc;
+    private Long acc;
 }
